Fix file tokenisation
The last word in a file was sometimes skipped, due to the file not ending with a newline character or some other recognised word end character.
This meant some checks would fail, or would ignore problematic areas of code at the end of the file.
Ensure the last word is included in all cases by ending the word after going through all characters in the file.