Support function prototypes with comments
If there are no comments found when checking for comments for a
function, check to see if there was a function prototype in an included
file. If there is, and it has comments, do not add a comments error for
the function.

Has the side effect of making simpatico work with recursive includes,
due to only loading each included file once, and reusing the previously
loaded information if the same file is included elsewhere.